+++ src/components/Input.jsx
@@ -1,4 +1,27 @@
-const Input = ({ clickAddButtonHandler, todo, setValue }) => {
+const Input = ({ todo, setTodoList, todoList, setTodo }) => {
+  const setValue = (event) => {
+    const { name, value } = event.target;
+
+    setTodo({ ...todo, [name]: value });
+  };
+
+  const clickAddButtonHandler = (event) => {
+    event.preventDefault();
+    const newTodoList = {
+      // id: todoList[todoList.length - 1]?.id + 1 || 1,
+      id: Date.now(),
+      ...todo,
+      isDone: false,
+    };
+
+    setTodoList([...todoList, newTodoList]);
+    setTodo({ title: "", content: "" });
+    localStorage.setItem(
+      "todoItem",
+      JSON.stringify([...todoList, newTodoList])
+    );
+  };
+
   return (
     <form className="add-todo" onSubmit={clickAddButtonHandler}>
       <div className="input-group">
